+++ src/containers/BankApi/BankApiInvalidBlocks/index.tsx
@@ -1,8 +1,7 @@
 import React, {FC} from 'react';
 import {NavLink} from 'react-router-dom';
 
-import {DocContainer, DocEndpoint, RequestResponseSnippet, TableParams} from 'components';
-import {QueryParamsOffsetAndLimit} from 'components/QueryParams';
+import {DocContainer, DocEndpoint, QueryParamsOffsetAndLimit, RequestResponseSnippet, TableParams} from 'components';
 
 const BankApiInvalidBlocks: FC = () => {
   return (
